use write error util function
