+++ internal/jimmhttp/auth_handler.go
@@ -231,18 +231,3 @@ func (oah *OAuthHandler) Whoami(w http.ResponseWriter, r *http.Request) {
 		zapctx.Error(ctx, "failed to write whoami body", zap.Error(err))
 	}
 }
-
-// writeError writes an error and logs the message. It is expected that the status code
-// is an erroneous status code.
-func writeError(ctx context.Context, w http.ResponseWriter, status int, err error, logMessage string) {
-	zapctx.Error(ctx, logMessage, zap.Error(err))
-	w.WriteHeader(status)
-	errMsg := ""
-	if err != nil {
-		errMsg = " - " + err.Error()
-	}
-	_, err = w.Write([]byte(http.StatusText(status) + errMsg))
-	if err != nil {
-		zapctx.Error(ctx, "failed to write status text error", zap.Error(err))
-	}
-}
